Migrate NodeClaim tests from Upstream repo
