--- test/suites/nodeclaim/testdata/al2023_userdata_custom_labels_input.yaml
@@ -1,15 +1,0 @@
-apiVersion: node.eks.aws/v1alpha1
-kind: NodeConfig
-spec:
-  cluster:
-    name: %s
-    apiServerEndpoint: %s
-    certificateAuthority: %s
-    cidr: 10.100.0.0/16
-  kubelet:
-    config:
-      clusterDNS:
-      - 10.0.100.10
-    flags:
-    - --node-labels='testing/cluster=unspecified,custom-label=custom-value,custom-label2=custom-value2'
-    - --register-with-taints='karpenter.sh/unregistered:NoExecute'
